Mqtt2InfluxDB/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile pyproject.toml
#
aiocsv==1.2.4
    # via influxdb-client
aiohttp==3.8.5
    # via influxdb-client
aiokafka==0.8.1
    # via mittaridatapumppu-persister (pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
anyio==4.0.0
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   aiokafka
asyncio-mqtt==0.16.2
    # via mittaridatapumppu-persister (pyproject.toml)
attrs==23.1.0
    # via aiohttp
blinker==1.6.2
    # via flask
certifi==2023.7.22
    # via
    #   fvhiot
    #   httpcore
    #   httpx
    #   influxdb-client
    #   requests
    #   sentry-sdk
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via flask
configparser==6.0.0
    # via mittaridatapumppu-persister (pyproject.toml)
flask==2.3.3
    # via fvhiot
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fvhiot @ https://github.com/ForumViriumHelsinki/FVHIoT-python/archive/refs/tags/v0.3.1.zip
    # via mittaridatapumppu-persister (pyproject.toml)
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via
    #   mittaridatapumppu-persister (pyproject.toml)
    #   python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
influxdb==5.3.1
    # via mittaridatapumppu-persister (pyproject.toml)
influxdb-client[async,extras]==1.37.0
    # via mittaridatapumppu-persister (pyproject.toml)
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kafka-python==2.0.2
    # via
    #   aiokafka
    #   fvhiot
    #   mittaridatapumppu-persister (pyproject.toml)
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.5
    # via
    #   fvhiot
    #   influxdb
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.1
    # via aiokafka
paho-mqtt==1.6.1
    # via
    #   asyncio-mqtt
    #   mittaridatapumppu-persister (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   influxdb
    #   influxdb-client
python-telegram-bot==20.5
    # via mittaridatapumppu-persister (pyproject.toml)
pytz==2023.3.post1
    # via influxdb
reactivex==4.0.4
    # via
    #   influxdb-client
    #   ruuvitag-sensor
requests==2.31.0
    # via
    #   influxdb
    #   mittaridatapumppu-persister (pyproject.toml)
ruuvitag-sensor==2.2.0
    # via mittaridatapumppu-persister (pyproject.toml)
sentry-asgi==0.2.0
    # via mittaridatapumppu-persister (pyproject.toml)
sentry-sdk==1.30.0
    # via sentry-asgi
six==1.16.0
    # via
    #   influxdb
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.31.1
    # via fvhiot
typing-extensions==4.7.1
    # via reactivex
urllib3==2.0.4
    # via
    #   influxdb-client
    #   requests
    #   sentry-sdk
werkzeug==2.3.7
    # via flask
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
